WIP: Implement pagination for transactions page

+++ src/stores/transaction.rs
@@ -36,9 +36,9 @@ pub struct TransactionQuery {
     /// Include transactions within `date_range` (inclusive).
     pub date_range: Option<RangeInclusive<Date>>,
     /// Selects up to the first N (`limit`) transactions.
-    pub limit: Option<u64>,
+    pub limit: Option<u32>,
     /// Ignore the first N transactions. Only has an effect if `limit` is not `None`.
-    pub offset: u64,
+    pub offset: u32,
     /// Orders transactions by date in the order `sort_date`. None returns transactions in the
     /// order they are stored.
     pub sort_date: Option<SortOrder>,
